+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Add special case for dlsym(RTLD_NEXT) (#82)
 * Remove input events from window event masks (#365)
 * Add basic EGL support
+* Add commit version and date to window title
 
 ### Changed